+++ lib/basic/widgets.dart
@@ -1,0 +1,14 @@
+export 'widget.dart';
+export 'command.dart';
+export 'pack.dart';
+export 'group.dart';
+export 'scoreboard.dart';
+export 'score.dart';
+export 'text_components.dart';
+export 'file.dart';
+export 'extend.dart';
+export 'for_list.dart';
+export 'entity.dart';
+export 'item.dart';
+export 'block.dart';
+export 'location.dart';