+++ lib/core.dart
@@ -4,18 +4,5 @@ export 'build/create_project.dart';
 export 'build/build.dart';
 
 // basic widgets
-export 'basic/widget.dart';
-export 'basic/command.dart';
-export 'basic/pack.dart';
-export 'basic/group.dart';
-export 'basic/scoreboard.dart';
-export 'basic/score.dart';
-export 'basic/text_components.dart';
-export 'basic/file.dart';
-export 'basic/extend.dart';
-export 'basic/for_list.dart';
+export 'basic/widgets.dart';
 export 'objD/widgets.dart';
-export 'basic/entity.dart';
-export 'basic/item.dart';
-export 'basic/block.dart';
-export 'basic/location.dart';